renamed TrajectoryLogging to DynamicTrace
